+++ src/utils.ts
@@ -1,6 +1,8 @@
-import { omit } from 'lodash';
 import { ITicket } from './renderer/hooks/useSupabase';
 import { IHistoryItem, IProduct, PayMethod } from './types';
+
+export const MIN_DATE = new Date(0);
+export const MAX_DATE = new Date(99999999999999);
 
 export const today = () => {
   const dateFormat = new Intl.DateTimeFormat('es-AR', {
@@ -215,6 +217,30 @@ export function downloadHistoryCSV(history: IHistoryItem[]): void {
   });
   downloadCSV(exportable);
 }
+
+export function downloadHistoryWithDetailCSV(history: IHistoryItem[]): void {
+  const exportable = history.flatMap((r) => {
+    return r.ticketLines.map((line) => {
+      return {
+        'Ticket Nro': r.id,
+        Fecha: new Date(r.date).toLocaleDateString('es-AR'),
+        Hora: new Date(r.date).toLocaleTimeString('es-AR'),
+        'Metodo de Pago': r.payMethod,
+        'Ticket Subtotal': toLocaleNumber(r.subTotal),
+        Descuento: toPercentage(r.discount),
+        'Ticket Total': toLocaleNumber(r.total),
+        Codigo: line.product.id,
+        Descripcion: line.product.descripcion,
+        'Precio Unitario': toLocaleNumber(line.product.precio),
+        'Precio Tarjeta': toLocaleNumber(line.product.precioTarjeta),
+        Cantidad: line.quantity,
+        'Subtotal Linea': toLocaleNumber(line.product.precio * line.quantity),
+      };
+    });
+  });
+  downloadCSV(exportable);
+}
+
 /**
  * Export data to a CSV file
  *
@@ -265,3 +291,11 @@ function convertArrayOfObjectsToCSV(array: any[]): string {
 
   return result;
 }
+
+export const debounce = (func: (...args: any[]) => any, wait: number) => {
+  let timeout: NodeJS.Timeout;
+  return (...args: any[]) => {
+    clearTimeout(timeout);
+    timeout = setTimeout(() => func(...args), wait);
+  };
+};